fix small unnecessary casts or references

+++ provers/stark/src/stark_prover.rs
@@ -345,9 +345,9 @@ impl<F: AbstractField> StarkProver<F> {
             transcript.add_elements(root, self.n_field_elements);
         } else {
             let hash: Vec<F> = vec![F::zero(); self.n_field_elements];
-            let n_hash = ((1 << (self.stark_info.as_ref().unwrap().stark_struct.steps[n_steps - 1].n_bits))
-                * Self::FIELD_EXTENSION as u64) as u64;
-            calculate_hash_c(p_stark, hash.as_ptr() as *mut c_void, fri_pol, n_hash as u64);
+            let n_hash = (1 << (self.stark_info.as_ref().unwrap().stark_struct.steps[n_steps - 1].n_bits))
+                * Self::FIELD_EXTENSION as u64;
+            calculate_hash_c(p_stark, hash.as_ptr() as *mut c_void, fri_pol, n_hash);
             transcript.add_elements(hash.as_ptr() as *mut c_void, self.n_field_elements);
         }
     }
